[added] web explanation + reduced ui modal cpu intensity
