--- lib/modals/backup/bookmark.dart
@@ -1,99 +1,0 @@
-import 'package:citizenwallet/theme/colors.dart';
-import 'package:citizenwallet/widgets/header.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter_svg/svg.dart';
-import 'package:go_router/go_router.dart';
-import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
-
-class BookmarkModal extends StatelessWidget {
-  final String title;
-
-  const BookmarkModal({
-    Key? key,
-    this.title = 'Wallet',
-  }) : super(key: key);
-
-  void handleDismiss(BuildContext context) {
-    FocusManager.instance.primaryFocus?.unfocus();
-    GoRouter.of(context).pop();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () => FocusManager.instance.primaryFocus?.unfocus(),
-      child: CupertinoPageScaffold(
-        backgroundColor: ThemeColors.uiBackgroundAlt.resolveFrom(context),
-        child: SafeArea(
-          minimum: const EdgeInsets.only(left: 10, right: 10, top: 20),
-          child: Flex(
-            direction: Axis.vertical,
-            children: [
-              Header(
-                title: title,
-                actionButton: CupertinoButton(
-                  padding: const EdgeInsets.all(5),
-                  onPressed: () => handleDismiss(context),
-                  child: Icon(
-                    CupertinoIcons.xmark,
-                    color: ThemeColors.touchable.resolveFrom(context),
-                  ),
-                ),
-              ),
-              Expanded(
-                child: Padding(
-                  padding: const EdgeInsets.fromLTRB(10, 0, 10, 0),
-                  child: CustomScrollView(
-                    controller: ModalScrollController.of(context),
-                    scrollBehavior: const CupertinoScrollBehavior(),
-                    physics: const AlwaysScrollableScrollPhysics(),
-                    slivers: [
-                      SliverFillRemaining(
-                        child: Column(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          children: [
-                            SvgPicture.asset(
-                              'assets/icons/bookmark_color.svg',
-                              semanticsLabel: 'voucher icon',
-                              height: 200,
-                              width: 200,
-                            ),
-                            const SizedBox(height: 100),
-                            Text(
-                              'Your wallet is in the link of the page you are on now.',
-                              textAlign: TextAlign.center,
-                              style: TextStyle(
-                                fontSize: 20,
-                                fontWeight: FontWeight.normal,
-                                color: ThemeColors.text.resolveFrom(context),
-                              ),
-                            ),
-                            const SizedBox(
-                              height: 20,
-                            ),
-                            const Text(
-                              "Bookmark this page to save your wallet.",
-                              textAlign: TextAlign.center,
-                              style: TextStyle(
-                                fontSize: 20,
-                                fontWeight: FontWeight.normal,
-                              ),
-                            ),
-                            const SizedBox(
-                              height: 140,
-                            ),
-                          ],
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
